remove superfulous component
